% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang_B.R
\name{Erlang_B}
\alias{Erlang_B}
\title{Erlang B Formula (aka. Erlang loss Formula)}
\usage{
Erlang_B(c, lambda, mu)
}
\arguments{
\item{c}{number of servers}

\item{lambda}{arrival rate}

\item{mu}{service rate}
}
\value{
Probability
}
\description{
Returns the probability that an arriving customer is blocked.
}
