% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pncc.R
\name{Pncc}
\alias{Pncc}
\title{Pn Function}
\usage{
Pncc(c, mu, lambda, n)
}
\arguments{
\item{c}{number of server}

\item{mu}{service rate}

\item{lambda}{arrival rate}

\item{n}{number of subjects/objects in the system}
}
\value{
Probability
}
\description{
Probability Pn of having n customers in the System
}
